% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{as.matrix.powerSimListCI}
\alias{as.matrix.powerSimListCI}
\alias{as.data.frame.powerSimListCI}
\title{Convenience functions for powerSimList confidence intervals}
\usage{
\method{as.matrix}{powerSimListCI}(x, ...)

\method{as.data.frame}{powerSimListCI}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a \code{powerSimList} object}

\item{...}{included for S3 compatibility

  This functions are simple convienence wrappers to flatten
  \code{confint.powerSimList} output, which is returned as a list of
  \code{powerSim} confidence intervals, just as if you had used \code{lapply}
  on a list of \code{powerSim} objects.}

\item{row.names}{see \code{\link{as.data.frame}}}

\item{optional}{see \code{\link{as.data.frame}}}
}
\description{
Manipulate and wrangle lists of confidence intervals into usable rectangular data formats
}
\seealso{
\code{\link{as.matrix}}

\code{\link{as.data.frame}}
}
