% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_half_life.R
\name{cost_half_life}
\alias{cost_half_life}
\title{Return the cost scale of a sim object}
\usage{
cost_half_life(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object derived from that frame}
}
\value{
The cost scale of the sim object
}
\description{
Return the cost scale of a sim object
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
cost_scale(mdl)
}
