% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_scale.R
\name{cost_scale}
\alias{cost_scale}
\title{Return the cost scale of a sim object}
\usage{
cost_scale(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object derived from that frame}
}
\value{
The cost scale of the sim object
}
\description{
Return the cost scale of a sim object
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
cost_scale(mdl)
}
