% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dest_coefs.R
\name{dest_coefs}
\alias{dest_coefs}
\title{Extract SIM model coefficients relating to destination}
\usage{
dest_coefs(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object}
}
\value{
A tibble with a list of coefficient values - either a balancing term for each area or
regression coefficients for the variables used to measure attractiveness of each area.
}
\description{
Extract SIM model coefficients relating to destination
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% doubly_constrained_sim(From,To,Dist,Count) -> mdl
mdl \%>\% dest_coefs
}
