% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dest_contrained_sim.R
\name{dest_constrained_sim}
\alias{dest_constrained_sim}
\title{Calibrate an origin-constrained spatial interaction model (SIM) via a Poisson regression.}
\usage{
dest_constrained_sim(df, orig, dest, prod, cost, intr)
}
\arguments{
\item{df}{A tibble or a data frame.}

\item{orig}{A column reference - the origin (a character or factor)}

\item{dest}{A column reference - the destination (a character or factor)}

\item{prod}{A data frame (or tibble) of origin production variables}

\item{cost}{A column reference - the cost (a character or factor)}

\item{intr}{A column reference - the interaction count (a character or factor)}

\item{type}{What type of model to use (unconstr, orig, dest, dbl)}
}
\value{
A 'sim' object - like a model object but with a few extra attributes.
}
\description{
Calibrate an origin-constrained spatial interaction model (SIM) via a Poisson regression.
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count)
}
