% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dest_coefs.R
\name{dest_scores}
\alias{dest_scores}
\title{Extract SIM model scores relating to destination}
\usage{
dest_scores(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object}
}
\value{
A tibble with a list of attractiveness scores - either based on balancing terms for
each area,  or on the basis of the attractiveness variables,  depending on whether destination constraints
were used.
}
\description{
Extract SIM model scores relating to destination
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% orig_constrained_sim(From,To,eastmid_empl,Dist,Count) -> mdl
mdl \%>\% dest_scores
}
