% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.R
\name{doFit}
\alias{doFit}
\title{Fit model to a new response.}
\usage{
doFit(y, model, subset)
}
\arguments{
\item{y}{new values for the response variable (vector or matrix depending on the model).}

\item{model}{a previously fitted model object.}

\item{subset}{boolean vector specifying how much of the data to use. If missing, the model is fit to all
    the data. This argument needs to be implemented for \code{\link{powerCurve}} to work.}
}
\value{
a fitted model object.
}
\description{
This is normally an internal function, but it can be overloaded to extend \code{simr} to other packages.
}

