% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doSim.R
\name{doSim}
\alias{doSim}
\title{Generate simulated response variables.}
\usage{
doSim(object, ...)
}
\arguments{
\item{object}{an object to simulare from, usually a fitted model.}
}
\value{
a vector containing simulated response values (or, for models  with a multivariate response such as
    binomial gl(m)m's, a matrix of simulated response values). Suitable as input for \code{\link{doFit}}.
}
\description{
This is normally an internal function, but it can be overloaded to extend \code{simr} to other packages.
}

