% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubly_constrained.R
\name{doubly_constrained_sim}
\alias{doubly_constrained_sim}
\title{Calibrate an doubly-constrained spatial interaction model (SIM) via a Poisson regression.}
\usage{
doubly_constrained_sim(df, orig, dest, cost, intr)
}
\arguments{
\item{df}{A tibble or a data frame.}

\item{orig}{A column reference - the origin (a character or factor)}

\item{dest}{A column reference - the destination (a character or factor)}

\item{cost}{A column reference - the cost (a character or factor)}

\item{intr}{A column reference - the interaction count (a character or factor)}
}
\value{
A 'sim' object - like a model object but with a few extra attributes.
}
\description{
Calibrate an doubly-constrained spatial interaction model (SIM) via a Poisson regression.
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% doubly_constrained_sim(From,To,Dist,Count)
}
