% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eastmid_ec_act}
\alias{eastmid_ec_act}
\title{East Midlands employer data - enterprises}
\format{A data frame with 40 rows and 2 variables:
\describe{
  \item{From}{Name of origin LA of trip to work}
  \item{EcAct}{Count of economically active residents}
}}
\source{
\url{https://www.nomisweb.co.uk/query/construct/summary.asp?mode=construct&version=0&dataset=556}
}
\usage{
eastmid_ec_act
}
\description{
A dataset containing counts of economically active residents for
local authority (LA) areas in the East Midlands (UK) taken
from the 2011 UK Census
}
\details{
Contains National Statistics data © Crown copyright and database right [2018]
}
\keyword{datasets}
