% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eastmid_empl}
\alias{eastmid_empl}
\title{East Midlands employer data - enterprises}
\format{A data frame with 40 rows and 5 variables:
\describe{
  \item{To}{Name of destimation LA of trip to work}
  \item{Micro}{Count of Micro enterprises (0 to 9)}
  \item{Small}{Count of Small enterprises (10 to 49)}
  \item{Medium}{Count of Medium enterprises (50 to 249)}
  \item{Large}{Count of Large enterprises (250+)}
}}
\source{
\url{https://www.nomisweb.co.uk/query/construct/summary.asp?mode=construct&version=0&dataset=142}
}
\usage{
eastmid_empl
}
\description{
A dataset containing counts of employers for
local authority (LA) areas in the East Midlands (UK) taken
from the Inter Departmental Business Register (IDBR)
}
\details{
Contains National Statistics data © Crown copyright and database right [2018]
}
\keyword{datasets}
