% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eastmid_sf}
\alias{eastmid_sf}
\title{East Midlands 'simple features' polygon object}
\format{An \code{sf} data frame with 40 rows and 3 variables:
\describe{
  \item{code}{Code for LA area}
  \item{name}{Name of LA area}
  \item{geometry}{Geometry column containing boundary shapes}
}}
\source{
\url{https://www.nomisweb.co.uk/query/construct/summary.asp?mode=construct&version=0&dataset=1211}
}
\usage{
eastmid_sf
}
\description{
An \code{sf} polygon data frame containing boundaries for
local authority (LA) areas in the East Midlands (UK).  A line
simplification algorithm (due to Visvalingham) maintaining
topological integrity has been applied.
}
\details{
Contains National Statistics data © Crown copyright and database right [2018] \cr
Contains OS data © Crown copyright [and database right] (2018) \cr
}
\references{
M. Visvalingam and J. D. Whyatt. Line generalisation by repeated elimination of points. The Cartographic Journal, 30(1):46–51, 1993.
}
\keyword{datasets}
