% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eastmid_ttw}
\alias{eastmid_ttw}
\title{East Midlands travel-to-work data}
\format{A data frame with 1600 rows and 4 variables:
\describe{
  \item{To}{Name of origin LA of trip to work}
  \item{From}{Name of destination LA of trip to work}
  \item{Dist}{Distance from origin to destimation (Crow flies)}
  \item{Count}{Number of people taking the \code{To} to \code{From} trip to work  }
}}
\source{
\url{https://www.nomisweb.co.uk/query/construct/summary.asp?mode=construct&version=0&dataset=1211}
}
\usage{
eastmid_ttw
}
\description{
A dataset contain ing travel-to-work counts for
local authority (LA) areas in the East Midlands (UK) taken
from the 2011 Census.
}
\details{
Contains National Statistics data © Crown copyright and database right [2018]
}
\keyword{datasets}
