% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simr.R
\docType{data}
\name{example}
\alias{example}
\title{Example dataset.}
\description{
A simple artificial data set used in the tutorial. There are two response variables,
a Poisson count \code{z} and a gaussian response \code{y}. There is a continuous predictor
\code{x} with ten levels \code{\{1,2,...,10\}} and a categorical predictor \code{g} with
three levels \code{\{a, b, c\}}.
}

