% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{getAllXnames}
\alias{getAllXnames}
\title{Extract all fixed-effect model terms}
\usage{
getAllXnames(fit, exclude = c("(Intercept)"), pred.type = c("anova", "coef"))
}
\arguments{
\item{fit}{a fitted model from \code{\link{lm}}, \code{\link{glm}},
\code{\link{lmer}},  \code{\link{glmer}}}

\item{exclude}{a vector of model terms to exclude, by default the intercept
term}

\item{pred.type}{type of predictor to extract.}
}
\description{
Extract all fixed-effect model terms
}
\details{
\code{pred.type} determines whether \code{anova}-type model "terms" are
  extracted or model coefficients (the actual $beta$'s). For numerical
  predictors, there is no difference. For categorical predictors,
  \code{pred.type="anova"} returns a single name per predictor, while
  \code{pred.type="coef"} returns a name for each contrast in the model. i.e.
  with categorical predictors unexpanded. \code{}
}
