% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get an object's data.}
\usage{
getData(object)
}
\arguments{
\item{a}{fitted model object (e.g. an object of class \code{merMod} or \code{lm}).}
}
\value{
A \code{data.frame} with the required data.
}
\description{
Get the data associated with a model object.
}
\details{
Looks for data in the following order:

\enumerate{
 \item{The object's \code{newData} attribute, if it has been set by \code{simr}.}
 \item{The \code{data} argument of \code{getCall(object)}, in the environment of \code{formula(object)}.}
}
}
\examples{
a <- lmer(Carbon ~ Year + (Year | Cluster), kiwifruit)
X <- getData(a)
}

