% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_scale.R
\name{interaction_scale}
\alias{interaction_scale}
\title{Return the interaction scale of a sim object}
\usage{
interaction_scale(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object derived from that frame}
}
\value{
The interaction scale of the sim object
}
\description{
Return the interaction scale of a sim object
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
interaction_scale(mdl)
}
