% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{kiwifruit}
\alias{kiwifruit}
\title{ARGOS kiwifruit soil monitoring pilot data}
\format{A data frame with 648 rows and 12 variables}
\usage{
data(kiwifruit)
}
\description{
ARGOS kiwifruit soil monitoring pilot data
}
\keyword{datasets}

