% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lastResult}
\alias{lastResult}
\title{Recover an unsaved simulation}
\usage{
lastResult()
}
\description{
Simulations can take a non-trivial time to run. If the user forgets to assign
the result to a variable this method can recover it.
}
\seealso{
\code{\link{.Last.value}}
}

