% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{makeGlmer}
\alias{makeGlmer}
\alias{makeLmer}
\title{Create an artificial mixed model object}
\usage{
makeGlmer(formula, family, fixef, VarCorr, data)

makeLmer(formula, fixef, VarCorr, sigma, data)
}
\arguments{
\item{formula}{a formula describing the model (see \code{\link[lme4]{glmer}}).}

\item{family}{type of response variable (see \code{\link{family}}).}

\item{fixef}{vector of fixed effects}

\item{VarCorr}{variance and covariances for random effects.
If there are multiple random effects, supply their parameters as a list.}

\item{data}{\code{data.frame} of explanatory variables.}

\item{sigma}{residual standard deviation.}
}
\description{
Make a \code{\link[lme4]{merMod}} object with the specified structure and parameters.
}
