% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modify.R
\name{Modifying parameters}
\alias{Modifying parameters}
\alias{VarCorr<-}
\alias{fixef<-}
\alias{sigma<-}
\title{Modify fixed effect coefs.}
\usage{
fixef(object) <- value

sigma(object) <- value

VarCorr(object) <- value
}
\arguments{
\item{object}{a linear mixed-effects model (\code{lmerMod}) object.}

\item{value}{a new vector of fixed effect coefficients.}
}
\description{
This function replaces the fixed effect coefficients in a fitted model.
}
\details{
This function would normally be used to change the value of individual fixed effect
coefficients, see the example for this usage.
}
\examples{
fm <- lmer(y ~ x + (1|g), data=example)
fixef(fm)
fixef(fm)["x"] <- -0.1
fixef(fm)
}

