% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orig_coefs.R
\name{orig_coefs}
\alias{orig_coefs}
\title{Create a tbl with the}
\usage{
orig_coefs(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object derived from that frame}

\item{df}{A data frame}

\item{colname}{A reference to the name of the new column - defaults to 'fitted'}
}
\value{
The data frame with an extra column for the fitted value
}
\description{
Create a tbl with the
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
eastmid_ttw \%>\% add_sim_fitted(mdl)
}
