% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orig_coefs.R
\name{orig_scores}
\alias{orig_scores}
\title{Extract SIM model scores relating to origin}
\usage{
orig_scores(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object}
}
\value{
A tibble with a list of productiveness scores - either based on balancing terms for
each area,  or on the basis of the productiveness variables,  depending on whether origin constraints
were used.
}
\description{
Extract SIM model scores relating to origin
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% dest_constrained_sim(From,To,eastmid_ec_act,Dist,Count) -> mdl
mdl \%>\% orig_scores
}
