% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson_reg}
\alias{poisson_reg}
\title{Carry out a Poisson regression.}
\usage{
poisson_reg(df, orig, dest, cost, intr, type = "dbl")
}
\arguments{
\item{df}{A tibble or a data frame.}

\item{orig}{A column reference - the origin (a character or factor)}

\item{dest}{A column reference - the destination (a character or factor)}

\item{cost}{A column reference - the cost (a character or factor)}

\item{intr}{A column reference - the interaction count (a character or factor)}

\item{type}{What type of model to use (unconstr, orig, dest, dbl)}
}
\value{
A 'sim' object - like a model object but with a few extra attributes.
}
\description{
Carry out a Poisson regression.
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count)
}
