
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerggplot.R
\name{powerggplot}
\alias{powerggplot}
\title{Create customizable plot of power curve based on ggplot2.}
\usage{
  powerggplot(x)
}
\arguments{
  \item{x}{a power curve (see \code{\link{powerCurve}}).}
}
\description{
  This function plots a \code{\link{powerCurve}} using \link[ggplot2], which then allows flexible modifications.
}
\seealso{
  \code{\link{powerPlot}}
}
