% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/powerSim.R
\name{powerSim}
\alias{powerSim}
\title{Calculate the power for an analysis.}
\usage{
powerSim(fit, test = fixed(getDefaultXname(fit)), sim = fit,
  nsim = getSimrOption("nsim"), alpha = 0.05, seed, ...)
}
\arguments{
\item{fit}{a linear mixed model object.}

\item{test}{the statistical test to perform, default is likelihood ratio.}

\item{sim}{an object to simulate from, by default this is the same as \code{fit}.}

\item{nsim}{the number of simulations to run.}

\item{alpha}{the significance level for the statistical test. Defaults to 0.05.}

\item{seed}{specify a random number generator seed, for reproducible results.}

\item{xname}{the name of the explanatory variable to be tested for significance.}
}
\description{
Performs a simulation analysis for a linear mixed model.
}
\examples{
fm1 <- lmer(y ~ x + (1|g), data=example)
powerSim(fm1, nSim=10)
}

