% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sim}
\alias{print.sim}
\title{Print method for a 'sim' object}
\usage{
\method{print}{sim}(sim_obj, ...)
}
\arguments{
\item{sim_obj}{A 'sim' object}
}
\description{
Print method for a 'sim' object
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count)
}
