% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extend.R
\name{reduceFrame}
\alias{reduceFrame}
\title{Reduce a longitudinal model to a single time period.}
\usage{
reduceFrame(object, along, level = X[[along]][1])
}
\description{
Reduce a longitudinal model to a single time period.
}

