% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setSimrNSim}
\alias{setSimrNSim}
\title{Set default number of simulations}
\usage{
setSimrNSim(nSim)
}
\arguments{
\item{nSim}{new value for the default number of simulations.}
}
\description{
Set the default number of simulations for each power analysis.
}
\examples{
setSimrNSim(10)
}

