% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setSimrProgress}
\alias{setSimrProgress}
\title{Set the default progress bar behaviour}
\usage{
setSimrProgress(x)
}
\arguments{
\item{x}{\code{TRUE} or \code{FALSE} should simulations use a progress bar?}
}
\description{
Set the default progress bar behaviour
}

