% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_aic.R
\name{sim_aic}
\alias{sim_aic}
\title{Return the AIC of a SIM model}
\usage{
sim_aic(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object}
}
\value{
AIC of the SIM
}
\description{
Return the AIC of a SIM model
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
sim_aic(mdl)
}
