% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_scenario.R
\name{sim_scenario}
\alias{sim_scenario}
\title{Create a scenario for SIM modelling based on costs, production and attraction variables}
\usage{
sim_scenario(sim_obj, cost = NULL, production = NULL, attraction = NULL)
}
\arguments{
\item{sim_obj}{A SIM object}

\item{cost}{A cost data frame with origin, destination and cost columns}

\item{production}{A production data frame - origin column plus production variable columns}

\item{attraction}{A attraction data frame - destinmation column plus attraction variable columns}
}
\value{
A scenartio data frame merging the information above
}
\description{
Create a scenario for SIM modelling based on costs, production and attraction variables
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
eastmid_ttw \%>\% add_sim_fitted(mdl)
}
