% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_vars.R
\name{sim_vars}
\alias{sim_vars}
\title{Return the quosures of the SIM variables for a given sim object}
\usage{
sim_vars(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object}
}
\value{
A list of quosures
}
\description{
Return the quosures of the SIM variables for a given sim object
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
sim_vars(mdl)
}
