% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simr.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Example dataset.}
\description{
A simple artificial data set used in the tutorial. There are two response variables,
a Poisson count \code{z} and a Gaussian response \code{y}. There is a continuous predictor
\code{x} with ten values \code{\{1,2,...,10\}} and a categorical predictor \code{g} with
three levels \code{\{a, b, c\}}.
}
