% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simr.R
\docType{package}
\name{simr-package}
\alias{simr-package}
\title{simr: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

