% Generated by roxygen2 (4.0.0): do not edit by hand
\name{simrOptions}
\alias{getSimrOption}
\alias{simrOptions}
\title{Options Settings for \code{simr}}
\usage{
simrOptions(...)

getSimrOption(x)
}
\arguments{
\item{...}{}

\item{x}{option name (character string)}
}
\value{
\code{getSimrOption} returns the current value for the option \code{x}.

\code{simrOptions} returns

\enumerate{
\item a named list of all options, if no arguments are given.
\item a named list of specified options, if a list of option names is given.
\item (invisibly) a named list of changed options with their previous values, if options are set.
}
}
\description{
Control the default behaviour of \code{simr} analyses.
}
\section{Options in \code{simr}}{


There are options
}
\examples{
getSimrOption("nSim")
oldopts <- simrOptions(nSim=5)
getSimrOption("nSim")
simrOptions(oldopts)
getSimrOption("nSim")
}

