% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testLibrary.R
\name{tests}
\alias{compare}
\alias{fixed}
\alias{random}
\alias{tests}
\title{Specify a statistical test to apply}
\usage{
fixed(xname, method = c("lr", "z", "kr", "pb"))

compare(model, method = c("lr", "pb"))

random()
}
\arguments{
\item{xname}{}

\item{method}{}

\item{model}{}
}
\description{
Specify a statistical test to apply
}
\examples{
fm1 <- lmer(y ~ x + (x|g), data=example)
powerSim(fm1, compare(. ~ x + (1|g)), nsim=25)
}

