% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconstrained_sim.R
\name{unconstrained_sim}
\alias{unconstrained_sim}
\title{Calibrate an unconstrained spatial interaction model (SIM) via a Poisson regression.}
\usage{
unconstrained_sim(df, orig, dest, attr, prod, cost, intr)
}
\arguments{
\item{df}{A tibble or a data frame.}

\item{orig}{A column reference - the origin (a character or factor)}

\item{dest}{A column reference - the destination (a character or factor)}

\item{attr}{A data frame (or tibble) of destination attraction variables}

\item{prod}{A data frame (or tibble) of origin production variables}

\item{cost}{A column reference - the cost (for example, a distance, financial cost or travel time)}

\item{intr}{A column reference - the interaction count}

\item{type}{What type of model to use (unconstr, orig, dest, dbl)}
}
\value{
A 'sim' object - like a model object but with a few extra attributes.
}
\description{
Calibrate an unconstrained spatial interaction model (SIM) via a Poisson regression.
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% unconstrained_sim(From,To,Dist,Count)
}
