% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sim_fiited.R
\name{add_sim_fitted}
\alias{add_sim_fitted}
\title{Add a column with the predicted data from a model}
\usage{
add_sim_fitted(df, sim_obj, colname, scenario = NULL)
}
\arguments{
\item{df}{A data frame}

\item{sim_obj}{A 'sim' object derived from that frame}

\item{colname}{A reference to the name of the new column - defaults to 'fitted'}

\item{scenario}{A scenario for prediction.  If NULL fits are based on original data}
}
\value{
The data frame with an extra column for the fitted value
}
\description{
Add a column with the predicted data from a model
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% poisson_reg(From,To,Dist,Count) -> mdl
eastmid_ttw \%>\% add_sim_fitted(mdl)
}
