% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTest.R
\name{doTest}
\alias{doTest}
\title{Apply a hypothesis test to a fitted model.}
\usage{
doTest(object, test, ...)
}
\arguments{
\item{object}{an object to apply a statistcal test to, usually a fitted model.}

\item{test}{a test function, see \link{tests}.}

\item{...}{additional options.}
}
\value{
a p-value with attributes describing the test.
}
\description{
This is normally an internal function, but it can be overloaded to extend \code{simr} to other packages.
}

