% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simr.R
\docType{data}
\name{kiwifruit}
\alias{kiwifruit}
\title{ARGOS kiwifruit soil monitoring pilot data}
\format{A data frame with 648 rows and 12 variables}
\usage{
data(kiwifruit)
}
\description{
ARGOS kiwifruit soil monitoring pilot data
}
\keyword{datasets}

