% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oops.R
\name{lastResult}
\alias{lastResult}
\title{Recover an unsaved simulation}
\usage{
lastResult()
}
\description{
Simulations can take a non-trivial time to run. If the user forgets to assign
the result to a variable this method can recover it.
}
\examples{
lm1 <- lmer(y ~ x + (1|g), data=example)
powerSim(lm1, nsim=25)
ps1 <- lastResult()
}
\seealso{
\code{\link[base]{.Last.value}}
}

