% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oops.R
\name{lastResult}
\alias{lastResult}
\title{Recover an unsaved simulation}
\usage{
lastResult()
}
\description{
Simulations can take a non-trivial time to run. If the user forgets to assign
the result to a variable this method can recover it.
}
\examples{
fm1 <- lmer(y ~ x + (1|g), data=simdata)
powerSim(fm1, nsim=10)
ps1 <- lastResult()

}
\seealso{
\code{\link[base]{.Last.value}}
}
