% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrapper.R
\name{nullOpt}
\alias{nullOpt}
\title{'Fit' a mixed model by returning previously calculated parameters.}
\usage{
nullOpt(fn, par, lower, upper, control)
}
\description{
Dummy 'optimizer' that allows `lmer` to be called without re-estimating parameters.
}

