% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orig_coefs.R
\name{orig_coefs}
\alias{orig_coefs}
\title{Extract SIM model coefficients relating to origin}
\usage{
orig_coefs(sim_obj)
}
\arguments{
\item{sim_obj}{A 'sim' object}
}
\value{
A tibble with a list of coefficient values - either a balancing term for each area or
regression coefficients for the variables used to measure productiveness of each area.
}
\description{
Extract SIM model coefficients relating to origin
}
\examples{
require(tidyverse)
eastmid_ttw \%>\% doubly_constrained_sim(From,To,Dist,Count) -> mdl
mdl \%>\% orig_coefs
}
