% Generated by roxygen2 (4.0.0): do not edit by hand
\name{powerCurve}
\alias{powerCurve}
\title{Calculate the power for an analysis at a range of levels.}
\usage{
powerCurve(fit, nSim = .SIMRDEFAULT_NSIM, xname = getDefaultXname(fit),
  along = xname, sim = fit, pval = 0.05, seed = 23)
}
\description{
Calculate the power for an analysis at a range of levels.
}

