
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerggplot.R
\name{powerggplot}
\alias{powerggplot}
\title{Create customizable plot of power curve through ggplot2.}
\usage{
  powerCurve(x)
}
\arguments{
  \item{x}{a power curve (see \code{\link{powerCurve}}).}
}
\description{
  This function plots a \code{\link{powerCurve}} using ggplot2, which then allows aesthetic modifications.
}
\seealso{
  \code{\link{print.powerCurve}}, \code{\link{summary.powerCurve}}, \code{\link{confint.powerCurve}}
}
