% Generated by roxygen2 (4.0.0): do not edit by hand
\name{powerSim}
\alias{powerSim}
\title{Calculate the power for an analysis.}
\usage{
powerSim(fit, nSim = .simrOptions$nSim, sim = fit,
  xname = getDefaultXname(fit), test, alpha = 0.05, seed, ...)
}
\arguments{
\item{fit}{fit}

\item{nSim}{nsim}

\item{sim}{sim}

\item{xname}{}

\item{test}{}

\item{alpha}{}

\item{seed}{}
}
\description{
Performs a simulation analysis for a linear mixed model.
}

