% Generated by roxygen2 (4.0.0): do not edit by hand
\name{powerSim}
\alias{powerSim}
\title{Calculate the power for an analysis.}
\usage{
powerSim(fit, nSim = .SIMRDEFAULT_NSIM, sim = fit,
  xname = getDefaultXname(fit), test, seed = NA, ...)
}
\description{
Calculate the power for an analysis.
}

