% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/powerSim.R
\name{powerSim}
\alias{powerSim}
\title{Estimate power by simulation.}
\usage{
powerSim(fit, test = fixed(getDefaultXname(fit)), sim = fit,
  nsim = getSimrOption("nsim"), alpha = 0.05, seed, ...)
}
\arguments{
\item{fit}{a fitted model object (see \code{\link{doFit}}).}

\item{test}{specify the test to perform. By default, the first fixed effect in \code{fit} will be tested.
(see: \link{tests}).}

\item{sim}{an object to simulate from. By default this is the same as \code{fit} (see \code{\link{doSim}}).}

\item{nsim}{the number of simulations to run.}

\item{alpha}{the significance level for the statistical test. Defaults to 0.05.}

\item{seed}{specify a random number generator seed, for reproducible results.}

\item{...}{any additional arguments are passed on to \code{\link{doFit}}.}
}
\description{
Performs a simulation analysis for a mixed model.
}
\examples{
fm1 <- lmer(y ~ x + (1|g), data=example)
powerSim(fm1, nSim=10)
}

