% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simr.R
\docType{package}
\name{simr-package}
\alias{simr-package}
\title{simr: Simulation-based power calculations for mixed models.}
\description{
\code{simr} is a package that makes it easy to run simulation-based power analyses
with \code{lme4}.
}

