% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/options.R
\name{simrOptions}
\alias{getSimrOption}
\alias{simrOptions}
\title{Options Settings for \code{simr}}
\usage{
simrOptions(...)

getSimrOption(x)
}
\arguments{
\item{...}{a list of names to get options, or a named list of new values to set options.}

\item{x}{option name (character string).}
}
\value{
\code{getSimrOption} returns the current value for the option \code{x}.

\code{simrOptions} returns

\enumerate{
\item a named list of all options, if no arguments are given.
\item a named list of specified options, if a list of option names is given.
\item (invisibly) a named list of changed options with their previous values, if options are set.
}
}
\description{
Control the default behaviour of \code{simr} analyses.
}
\section{Options in \code{simr}}{


Options that can be set with this method (and their initial values).

\code{nSim}: default number of simulations (\code{1000}).
\code{progress}: use progress bars during calculations (\code{TRUE}).
\code{binom}: method for calculating confidence intervals (\code{"exact"}).
}
\examples{
getSimrOption("nsim")
oldopts <- simrOptions(nsim=5)
getSimrOption("nsim")
simrOptions(oldopts)
getSimrOption("nsim")
}

